fix intervention dto in creation when location type changed
