+++ client/src/containers/interventions/form/types.ts
@@ -1,0 +1,4 @@
+import type * as yup from 'yup';
+import type schemaValidation from './schema-validation';
+
+export type SubSchema = yup.InferType<typeof schemaValidation>;
